ADD: entities can store component can get component can remove component has Display system

+++ src/game/components/Position.cpp
@@ -12,31 +12,31 @@ Position::Position() :_x(0), _y(0)
   _name = "Position";
 }
 
-Position::Position(int x, int y) :_x(x), _y(y)
+Position::Position(float x, float y) :_x(x), _y(y)
 {
   _name = "Position";
 }
 
 //getters
 
-int Position::getX() const
+float Position::getX() const
 {
   return _x;
 }
 
-int Position::getY() const
+float Position::getY() const
 {
   return _y;
 }
 
 //setters
 
-void Position::setX(int x)
+void Position::setX(float x)
 {
   _x  = x;
 }
 
-void Position::setY(int y)
+void Position::setY(float y)
 {
   _y = y;
 }
